requirements.txt
bert-tensorflow
dopamine-rl==3.0.1
regex
tensor2tensor
tensorflow
keras_transformer
matplotlib
pandas
scikit-learn
javalang
progressbar2
